Correcting a typo on line #434
Line 434: "Good question. First off, if you want to update individual values in a list you're likely going to need..." 

The next word here after "need" is "to," but I believe it should be an "a," so I changed it to "a" (as in" "...a loop that includes the index values).